+++ src/components/table/core/tbody.ts
@@ -1,6 +1,6 @@
 import { customElement } from 'lit/decorators.js'
 
-import { ClassifiedElement } from '../classified-element.js'
+import { ClassifiedElement } from '../../classified-element.js'
 
 // tl;dr <tbody/>, table-row-group
 @customElement('astra-rowgroup')
