+++ src/components/table/core/thead.ts
@@ -1,6 +1,6 @@
 import { customElement } from 'lit/decorators.js'
 
-import { ClassifiedElement } from '../classified-element.js'
+import { ClassifiedElement } from '../../classified-element.js'
 
 // tl;dr <thead/>, table-header-group
 @customElement('astra-thead')
