+++ src/components/table/menu/index.ts
@@ -3,11 +3,11 @@ import { property, state } from 'lit/decorators.js'
 import { classMap } from 'lit/directives/class-map.js'
 import { repeat } from 'lit/directives/repeat.js'
 
-import { CaretDown } from '../../icons/caret-down.js'
-import classMapToClassName from '../../lib/class-map-to-class-name.js'
-import { MenuOpenEvent, MenuSelectedEvent } from '../../lib/events.js'
-import { Theme, type HeaderMenuOptions } from '../../types.js'
-import { ClassifiedElement } from '../classified-element.js'
+import { CaretDown } from '../../../icons/caret-down.js'
+import classMapToClassName from '../../../lib/class-map-to-class-name.js'
+import { MenuOpenEvent, MenuSelectedEvent } from '../../../lib/events.js'
+import { Theme, type HeaderMenuOptions } from '../../../types.js'
+import { ClassifiedElement } from '../../classified-element.js'
 
 export class Menu extends ClassifiedElement {
   protected override classMap() {
