+++ src/components/table/outerbase-table.ts
@@ -2,17 +2,17 @@ import { css, html, type PropertyValueMap } from 'lit'
 import { customElement, property, state } from 'lit/decorators.js'
 import { classMap } from 'lit/directives/class-map.js'
 import { isEqual } from 'lodash-es'
-import { ArrowsClockwise } from '../icons/arrows-clockwise.js'
-import { CaretDown } from '../icons/caret-down.js'
-import { CaretLeft } from '../icons/caret-left.js'
-import { CaretRight } from '../icons/caret-right.js'
-import { Table as TableIcon } from '../icons/table.js'
-import { diffObjects } from '../lib/diff-objects.js'
-import { normalizeKeys } from '../lib/normalize-object-keys.js'
-import stringifyWithoutNewLines from '../lib/stringify-without-new-lines.js'
-import type { APIResponse, CellUpdateEvent, Fields, MenuSelectedEvent, RowAsRecord, Rows, SourceSchema, Table } from '../types.js'
-import './button.js' // Ensure the button component is imported
-import AstraTable from './table/index.js'
+import { ArrowsClockwise } from '../../icons/arrows-clockwise.js'
+import { CaretDown } from '../../icons/caret-down.js'
+import { CaretLeft } from '../../icons/caret-left.js'
+import { CaretRight } from '../../icons/caret-right.js'
+import { Table as TableIcon } from '../../icons/table.js'
+import { diffObjects } from '../../lib/diff-objects.js'
+import { normalizeKeys } from '../../lib/normalize-object-keys.js'
+import stringifyWithoutNewLines from '../../lib/stringify-without-new-lines.js'
+import type { APIResponse, CellUpdateEvent, Fields, MenuSelectedEvent, RowAsRecord, Rows, SourceSchema, Table } from '../../types.js'
+import '../button.js' // Ensure the button component is imported
+import AstraTable from './core/index.js'
 
 // const OUTERBASE_API_DOMAIN = 'app.outerbase.com'
 const OUTERBASE_API_DOMAIN = 'app.dev.outerbase.com'
