+++ src/index.ts
@@ -1,10 +1,10 @@
 import Button from './components/button.js'
 import Card from './components/card.js'
 import Input from './components/input.js'
-import OuterbaseTable from './components/outerbase-table.js'
 import ScrollArea from './components/scroll-area.js'
 import Select from './components/select.js'
-import Table from './components/table/index.js'
+import Table from './components/table/core/index.js'
+import OuterbaseTable from './components/table/outerbase-table.js'
 import Text from './components/text.js'
 
 // export * as Types from './types'
