+++ src/pages/c/table.astro
@@ -1,6 +1,7 @@
 ---
+import AstraTable from '../../components/table/core'
+
 import { Code } from 'astro/components'
-import AstraTable from '../../components/table'
 import ComponentLayout from '../../layouts/Component.astro'
 import Layout from '../../layouts/Layout.astro'
